// sdk/src/generated/EnWalletV5.gen.ts
type RemainingBitsAndRefs = c.Slice

type StoreCallback<T> = (obj: T, b: c.Builder) => void
type LoadCallback<T> = (s: c.Slice) => T

export type CellRef<T> = {
    ref: T
}
// ...
type uint8 = bigint
type uint32 = bigint
type uint64 = bigint
type uint256 = bigint
// ...
export interface OutActionWithSendMessageOnly {
    readonly $: 'OutActionWithSendMessageOnly'
    prev: CellRef<OutActionWithSendMessageOnlyOrEmpty>
    sendMode: uint8
    outMsg: c.Cell
}

export const OutActionWithSendMessageOnly = {
    PREFIX: 0x0ec3c86d,

    create(args: {
        prev: CellRef<OutActionWithSendMessageOnlyOrEmpty>
        sendMode: uint8
        outMsg: c.Cell
    }): OutActionWithSendMessageOnly {
        return {
            $: 'OutActionWithSendMessageOnly',
            ...args
        }
    },
    fromSlice(s: c.Slice): OutActionWithSendMessageOnly {
        loadAndCheckPrefix32(s, 0x0ec3c86d, 'OutActionWithSendMessageOnly');
        return {
            $: 'OutActionWithSendMessageOnly',
            prev: loadCellRef<OutActionWithSendMessageOnlyOrEmpty>(s, OutActionWithSendMessageOnlyOrEmpty.fromSlice),
            sendMode: s.loadUintBig(8),
            outMsg: s.loadRef(),
        }
    },
    store(self: OutActionWithSendMessageOnly, b: c.Builder): void {
        b.storeUint(0x0ec3c86d, 32);
        storeCellRef<OutActionWithSendMessageOnlyOrEmpty>(self.prev, b, OutActionWithSendMessageOnlyOrEmpty.store);
        b.storeUint(self.sendMode, 8);
        b.storeRef(self.outMsg);
    },
    toCell(self: OutActionWithSendMessageOnly): c.Cell {
        return makeCellFrom<OutActionWithSendMessageOnly>(self, OutActionWithSendMessageOnly.store);
    }
}

/**
 > type OutActionWithSendMessageOnlyOrEmpty = OutActionWithSendMessageOnly | void
 */
export type OutActionWithSendMessageOnlyOrEmpty =
    | OutActionWithSendMessageOnly
    | { $: 'void', value: void }

export const OutActionWithSendMessageOnlyOrEmpty = {
    fromSlice(s: c.Slice): OutActionWithSendMessageOnlyOrEmpty {
        return (s.remainingBits === 0 && s.remainingRefs === 0) ? { $: 'void', value: void 0 } : OutActionWithSendMessageOnly.fromSlice(s);
    },
    store(self: OutActionWithSendMessageOnlyOrEmpty, b: c.Builder): void {
        switch (self.$) {
            case 'OutActionWithSendMessageOnly':
                OutActionWithSendMessageOnly.store(self, b);
                break;
            case 'void':
                break;
        }
    },
    toCell(self: OutActionWithSendMessageOnlyOrEmpty): c.Cell {
        return makeCellFrom<OutActionWithSendMessageOnlyOrEmpty>(self, OutActionWithSendMessageOnlyOrEmpty.store);
    }
}

/**
 > type SnakedExtraActions = RemainingBitsAndRefs
 */
export type SnakedExtraActions = RemainingBitsAndRefs

export const SnakedExtraActions = {
    fromSlice(s: c.Slice): SnakedExtraActions {
        return loadTolkRemaining(s);
    },
    store(self: SnakedExtraActions, b: c.Builder): void {
        storeTolkRemaining(self, b);
    },
    toCell(self: SnakedExtraActions): c.Cell {
        return makeCellFrom<SnakedExtraActions>(self, SnakedExtraActions.store);
    }
}
// ...
export interface InternalSignedRequest {
    readonly $: 'InternalSignedRequest'
    walletId: uint32
    validUntil: uint32
    seqno: uint32
    outActions: CellRef<OutActionWithSendMessageOnly> | null
    hasExtraActions: boolean
    extraActions: SnakedExtraActions
}

export const InternalSignedRequest = {
    PREFIX: 0x73696e74,

    create(args: {
        walletId: uint32
        validUntil: uint32
        seqno: uint32
        outActions: CellRef<OutActionWithSendMessageOnly> | null
        hasExtraActions: boolean
        extraActions: SnakedExtraActions
    }): InternalSignedRequest {
        return {
            $: 'InternalSignedRequest',
            ...args
        }
    },
    fromSlice(s: c.Slice): InternalSignedRequest {
        loadAndCheckPrefix32(s, 0x73696e74, 'InternalSignedRequest');
        return {
            $: 'InternalSignedRequest',
            walletId: s.loadUintBig(32),
            validUntil: s.loadUintBig(32),
            seqno: s.loadUintBig(32),
            outActions: s.loadBoolean() ? loadCellRef<OutActionWithSendMessageOnly>(s, OutActionWithSendMessageOnly.fromSlice) : null,
            hasExtraActions: s.loadBoolean(),
            extraActions: SnakedExtraActions.fromSlice(s),
        }
    },
    store(self: InternalSignedRequest, b: c.Builder): void {
        b.storeUint(0x73696e74, 32);
        b.storeUint(self.walletId, 32);
        b.storeUint(self.validUntil, 32);
        b.storeUint(self.seqno, 32);
        storeTolkNullable<CellRef<OutActionWithSendMessageOnly>>(self.outActions, b,
            (v,b) => storeCellRef<OutActionWithSendMessageOnly>(v, b, OutActionWithSendMessageOnly.store)
        );
        b.storeBit(self.hasExtraActions);
        SnakedExtraActions.store(self.extraActions, b);
    },
    toCell(self: InternalSignedRequest): c.Cell {
        return makeCellFrom<InternalSignedRequest>(self, InternalSignedRequest.store);
    }
}

/**
 > struct (0x7369676e) ExternalSignedRequest {
 >     walletId: uint32
 >     validUntil: uint32
 >     seqno: uint32
 >     outActions: OutActionsCell?
 >     hasExtraActions: bool
 >     extraActions: SnakedExtraActions
 > }
 */
export interface ExternalSignedRequest {
    readonly $: 'ExternalSignedRequest'
    walletId: uint32
    validUntil: uint32
    seqno: uint32
    outActions: CellRef<OutActionWithSendMessageOnly> | null
    hasExtraActions: boolean
    extraActions: SnakedExtraActions
}

export const ExternalSignedRequest = {
    PREFIX: 0x7369676e,

    create(args: {
        walletId: uint32
        validUntil: uint32
        seqno: uint32
        outActions: CellRef<OutActionWithSendMessageOnly> | null
        hasExtraActions: boolean
        extraActions: SnakedExtraActions
    }): ExternalSignedRequest {
        return {
            $: 'ExternalSignedRequest',
            ...args
        }
    },
    fromSlice(s: c.Slice): ExternalSignedRequest {
        loadAndCheckPrefix32(s, 0x7369676e, 'ExternalSignedRequest');
        return {
            $: 'ExternalSignedRequest',
            walletId: s.loadUintBig(32),
            validUntil: s.loadUintBig(32),
            seqno: s.loadUintBig(32),
            outActions: s.loadBoolean() ? loadCellRef<OutActionWithSendMessageOnly>(s, OutActionWithSendMessageOnly.fromSlice) : null,
            hasExtraActions: s.loadBoolean(),
            extraActions: SnakedExtraActions.fromSlice(s),
        }
    },
    store(self: ExternalSignedRequest, b: c.Builder): void {
        b.storeUint(0x7369676e, 32);
        b.storeUint(self.walletId, 32);
        b.storeUint(self.validUntil, 32);
        b.storeUint(self.seqno, 32);
        storeTolkNullable<CellRef<OutActionWithSendMessageOnly>>(self.outActions, b,
            (v,b) => storeCellRef<OutActionWithSendMessageOnly>(v, b, OutActionWithSendMessageOnly.store)
        );
        b.storeBit(self.hasExtraActions);
        SnakedExtraActions.store(self.extraActions, b);
    },
    toCell(self: ExternalSignedRequest): c.Cell {
        return makeCellFrom<ExternalSignedRequest>(self, ExternalSignedRequest.store);
    }
}

/**
 > type ExtensionsDict = map<uint256, bool>
 */
export type ExtensionsDict = c.Dictionary<uint256, boolean>

export const ExtensionsDict = {
    fromSlice(s: c.Slice): ExtensionsDict {
        return c.Dictionary.load<uint256, boolean>(c.Dictionary.Keys.BigUint(256), c.Dictionary.Values.Bool(), s);
    },
    store(self: ExtensionsDict, b: c.Builder): void {
        b.storeDict<uint256, boolean>(self, c.Dictionary.Keys.BigUint(256), c.Dictionary.Values.Bool());
    },
    toCell(self: ExtensionsDict): c.Cell {
        return makeCellFrom<ExtensionsDict>(self, ExtensionsDict.store);
    }
}

/**
 > struct Storage {
 >     isSignatureAllowed: bool
 >     seqno: uint32
 >     subwalletId: uint32
 >     publicKey: uint256
 >     extensions: ExtensionsDict
 > }
 */
export interface Storage {
    readonly $: 'Storage'
    isSignatureAllowed: boolean
    seqno: uint32
    subwalletId: uint32
    publicKey: uint256
    extensions: ExtensionsDict
}

export const Storage = {
    create(args: {
        isSignatureAllowed: boolean
        seqno: uint32
        subwalletId: uint32
        publicKey: uint256
        extensions: ExtensionsDict
    }): Storage {
        return {
            $: 'Storage',
            ...args
        }
    },
    fromSlice(s: c.Slice): Storage {
        return {
            $: 'Storage',
            isSignatureAllowed: s.loadBoolean(),
            seqno: s.loadUintBig(32),
            subwalletId: s.loadUintBig(32),
            publicKey: s.loadUintBig(256),
            extensions: ExtensionsDict.fromSlice(s),
        }
    },
    store(self: Storage, b: c.Builder): void {
        b.storeBit(self.isSignatureAllowed);
        b.storeUint(self.seqno, 32);
        b.storeUint(self.subwalletId, 32);
        b.storeUint(self.publicKey, 256);
        ExtensionsDict.store(self.extensions, b);
    },
    toCell(self: Storage): c.Cell {
        return makeCellFrom<Storage>(self, Storage.store);
    }
}